+++ doc/changelog.rst
@@ -73,6 +73,7 @@ Localization
 Admin
 ~~~~~
 
+- Add PersonContact choice widget
 - Add barcode filter to product list view
 - Show stock in order creator
 - Fix bug: Fix decimal precision issues on order creation